refactor: reservation/payment 테스트코드 일괄 수정(진행중)
