+++ app/src/test/java/com/hotelJava/reservation/application/port/in/ConfirmUseCaseTest.java
@@ -1,0 +1,3 @@
+package com.hotelJava.reservation.application.port.in;
+
+class ConfirmUseCaseTest {}
