+++ app/src/test/java/com/hotelJava/reservation/domain/ReservationTest.java
@@ -1,0 +1,7 @@
+package com.hotelJava.reservation.domain;
+
+class ReservationTest {
+  void consumeStock() {}
+
+  void restoreStock() {}
+}
