+++ app/src/test/java/com/hotelJava/reservation/domain/ReservationValidatorTest.java
@@ -1,0 +1,3 @@
+package com.hotelJava.reservation.domain;
+
+class ReservationValidatorTest {}
