+++ app/src/test/java/com/hotelJava/reservation/service/LazyReservationServiceTest.java
@@ -1,0 +1,12 @@
+package com.hotelJava.reservation.service;
+
+import org.junit.jupiter.api.Test;
+
+class LazyReservationServiceTest {
+
+  @Test
+  void reserve() {}
+
+  @Test
+  void confirm() {}
+}
