+++ app/src/test/java/com/hotelJava/reservation/service/ReservationServiceManagerTest.java
@@ -1,0 +1,3 @@
+package com.hotelJava.reservation.service;
+
+class ReservationServiceManagerTest {}
